refactor and add badge for news

+++ app/src/main/java/de/nicidienase/geniesser_app/GourmetActivityViewModel.kt
@@ -9,6 +9,6 @@ class GourmetActivityViewModel(
     private val preferencesService: PreferencesService
 ) : ViewModel() {
 
-    val newsCount: LiveData<Long>
-            get() = newsRepository.newsCountForLocation(preferencesService.currentLocation.toLong())
+    val newsCount: LiveData<Int>
+            get() = newsRepository.newsCountForLocation(preferencesService.currentLocation)
 }
